refactor: 배포 중 에러로 인한 폴더 이름 복구
