+++ src/components/mypage/card-style/index.tsx
@@ -7,9 +7,9 @@ import PaginationBtn from '@/components/shared/PaginationBtn';
 import useFilterStore from '@/store/useFilterStore';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 
-import { MeetingCard } from './MeetingCard';
+import { MeetingCard } from './meeting-card';
 import ReviewCategory from '../category/ReviewCategory';
-import MyReviewList from '../MyReviewList';
+import MyReviewList from '../my-review-list';
 
 import Empty from 'public/lottie/empty.json';
