+++ src/components/mypage/card-style/reviewable-card/index.tsx
